+++ src/lib.rs
@@ -91,32 +91,50 @@
 //!     .initialize(terminal_path.as_str())
 //!     .expect("Unable to connect to terminal");
 //!
-//! let current_symbol = connection.symbol_info("EURUSD").unwrap();
+//! let current_symbol = connection.symbol_info("BTCUSD").unwrap();
+//!
+//! let symbol_name = current_symbol
+//!   .get_info_string(InfoProperties::SymbolInfoProperty(SymbolInfoProperty::Name))
+//!   .unwrap();
+//!
+//! let ask_price = current_symbol
+//!   .get_info_float(InfoProperties::SymbolInfoProperty(SymbolInfoProperty::Ask))
+//!   .unwrap();
+//!
+//! let bid_price = current_symbol
+//!   .get_info_float(InfoProperties::SymbolInfoProperty(SymbolInfoProperty::Bid))
+//!   .unwrap();
+//!
+//! let point = current_symbol
+//!   .get_info_float(InfoProperties::SymbolInfoProperty(
+//!     SymbolInfoProperty::Point,
+//!   ))
+//!   .unwrap();
 //!
 //! let open_trade = TradeRequestBuilder::new()
-//!     .action(TradeActionRequest::DEAL)
-//!     .symbol(current_symbol.name.clone())
-//!     .volume(0.01)
-//!     .price(current_symbol.ask)
-//!     .r#type(OrderType::BUY)
-//!     .type_filling(OrderTypeFilling::IOC)
-//!     .type_time(OrderTypeTime::GTC)
-//!     .sl(current_symbol.bid - (100.0 * current_symbol.point))
-//!     .tp(current_symbol.bid + (100.0 * current_symbol.point))
-//!     .comment("Test".to_string());
+//!   .action(TradeActionRequest::DEAL)
+//!   .symbol(symbol_name.clone())
+//!   .volume(0.01)
+//!   .price(ask_price)
+//!   .r#type(OrderType::BUY)
+//!   .type_filling(OrderTypeFilling::IOC)
+//!   .type_time(OrderTypeTime::GTC)
+//!   .sl(bid_price - (100.0 * point))
+//!   .tp(bid_price + (100.0 * point))
+//!   .comment("Test".to_string());
 //!
 //! let open_order = connection.order_send(open_trade).unwrap();
 //!
 //! let close_request = TradeRequestBuilder::new()
-//!     .action(TradeActionRequest::DEAL)
-//!     .symbol(current_symbol.name)
-//!     .volume(0.01)
-//!     .price(current_symbol.bid)
-//!     .position(open_order.order)
-//!     .r#type(OrderType::SELL)
-//!     .type_filling(OrderTypeFilling::IOC)
-//!     .type_time(OrderTypeTime::GTC)
-//!     .comment("Test".to_string());
+//!   .action(TradeActionRequest::DEAL)
+//!   .symbol(symbol_name)
+//!   .volume(0.01)
+//!   .price(bid_price)
+//!   .position(open_order.order)
+//!   .r#type(OrderType::SELL)
+//!   .type_filling(OrderTypeFilling::IOC)
+//!   .type_time(OrderTypeTime::GTC)
+//!   .comment("Test".to_string());
 //!
 //! let close_send = connection.order_send(close_request);
 //!