+++ src/prelude.rs
@@ -1,5 +1,5 @@
 pub type MQLResult<T> = Result<T, MQLError>;
-pub type MQLError = (i64, String);
+pub type MQLError = (RuntimeError, String);
 pub use crate::connection::python::MT5PythonConnection;
 pub use crate::enums::*;
 pub use crate::schemas::*;