+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased 0.1.2] - 2024-07-22
+## [Unreleased 0.2.0] - 2024-07-22
 
-## Fixed
+### Fixed
 - SymbolInfo data types now matches with MQL5 Symbol Properties
+
+### Changed
+- Major change on how to access AccountInfo, TerminalInfo and SymbolInfo
+
+### Added
+- Added RuntimeError enum for managing runtime error code.
 
 ## [Unreleased 0.1.1] - 2024-07-21